requirements.txt
# AWS SDK and Services
boto3==1.28.85
awscli==1.29.85
sagemaker==2.189.0

# Data Processing
pandas==2.1.1
numpy==1.25.2
pyarrow==13.0.0

# Machine Learning
scikit-learn==1.3.1
xgboost==2.0.0
imbalanced-learn==0.11.0

# Glue Development
awsglue-local==1.0.0
pyspark==3.4.1

# Data Validation
great-expectations==0.18.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.21
redshift-connector==2.0.918

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2

# Monitoring and Logging
cloudwatch-metrics==0.1.1

# Development
pytest==7.4.2
black==23.9.1
flake8==6.1.0
jupyter==1.0.0
notebook==7.0.4
